+++ jax_md/_nn/mace/__init__.py
@@ -1,0 +1,1 @@
+"""MACE modules for JAX-MD."""
