--- jax_md/_nn/mace_jax_interface/__init__.py
@@ -1,6 +1,0 @@
-"""MACE-JAX modules for JAX-MD.
-
-This package intentionally does not import its submodules. Some conversion and
-model-building helpers require optional MACE-JAX, MACE, or Torch dependencies;
-import the specific submodule needed at the call site.
-"""
